src/pi/requirements.txt
boto3>=1.34.0
pyyaml>=6.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
